mortgage-processor/requirements.txt
# FastAPI for document management API
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6  # For file uploads
pydantic==2.8.2
PyYAML==6.0.2
llama-stack-client
instructor
fire
sqlalchemy==2.0.23
alembic==1.12.1
# FastAPI auth dependencies removed - using LangGraph dev
typing-extensions
# Jinja2 removed - no longer using FastAPI templates

# LangGraph and LangChain dependencies
langgraph>=0.1.19
langchain>=0.1.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-experimental>=0.0.60  # For LLMGraphTransformer
langgraph-checkpoint-sqlite
aiosqlite  # Required for AsyncSqliteSaver

# Document processing dependencies
PyPDF2>=3.0.0  # PDF text extraction
python-docx>=1.1.0  # Word document processing
Pillow>=10.0.0  # Image processing
pytesseract>=0.3.10  # OCR for image text extraction

# Neo4j integration for graph database
neo4j>=5.0.0
langchain-neo4j>=0.0.6

# Using built-in LangGraph supervisor functionality - no external package needed

# LangGraph CLI for development server
langgraph-cli[inmem]>=0.1.19

# A2A (Agent-to-Agent) SDK for external agent integration
a2a>=0.1.0
httpx>=0.25.0  # For external agent HTTP calls